requirements.txt
flask==2.0.1
pandas==1.2.5
pillow==8.2.0
python-docx==0.8.11
python-dotenv==0.18.0
pycodestyle==2.7.0